feat: #41442 订单接口增加返回订单图片
